# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AProfile": {
        "description": "Profili AI selezionabili manualmente dall'utente.",
        "enum": [
          "assistant",
          "researcher",
          "coder"
        ],
        "title": "AProfile",
        "type": "string"
      },
      "BindingDTO": {
        "description": "Binding del profilo a runtime e modello, con parametri espliciti.",
        "properties": {
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "runtime": {
            "title": "Runtime",
            "type": "string"
          }
        },
        "required": [
          "name",
          "runtime",
          "model_name",
          "parameters"
        ],
        "title": "BindingDTO",
        "type": "object"
      },
      "BootstrapRequest": {
        "description": "Richiesta di bootstrap monouso: nome del proprietario + credenziale.\n\nLa credenziale non viene mai persistita in chiaro: il caso d'uso ne\ncalcola il record hash con ``newray.kernel.crypto`` (B-03.2-14). La\nlunghezza minima applicativa è verificata dal kernel; qui il limite\nsuperiore protegge da payload gonfiati senza fare da oracolo\nsull'algoritmo di hashing.",
        "properties": {
          "credential": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Credential",
            "type": "string"
          },
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "credential"
        ],
        "title": "BootstrapRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdentityDTO": {
        "description": "Identità pubblica del principal risolto dal server.\n\nNiente campi sensibili oltre il minimo necessario alla WebUI:\nnessun token, nessun dettaglio della sessione.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "display_name",
          "role"
        ],
        "title": "IdentityDTO",
        "type": "object"
      },
      "LoginRequest": {
        "description": "Richiesta di login del proprietario esistente (B-03.2-14).\n\nOwner sconosciuto e credenziale sbagliata collassano in un unico\n``INVALID_CREDENTIALS``: il canale non è un oracolo per scoprire il\nnome scelto in bootstrap.",
        "properties": {
          "credential": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Credential",
            "type": "string"
          },
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "credential"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "ModelInfoDTO": {
        "description": "Modello del runtime con digest e stato (NewRay.md §9.1).\n\nIl digest identifica l'artefatto per la riproducibilità; i tag\nmobili restano scoperta, mai riferimento degli snapshot.",
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digest"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "runtime": {
            "title": "Runtime",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ModelStatus"
          }
        },
        "required": [
          "name",
          "runtime",
          "digest",
          "status",
          "capabilities"
        ],
        "title": "ModelInfoDTO",
        "type": "object"
      },
...
      "ModelStatus": {
        "description": "Stati distinti del catalogo (NewRay.md §9.1): essere installati,\ncompatibili e qualificati sono fatti diversi.",
        "enum": [
          "installed",
          "compatible",
          "qualified"
        ],
        "title": "ModelStatus",
        "type": "string"
      },
      "ProfileDTO": {
        "description": "Profilo con versione corrente, binding e disponibilità del modello.\n\n``model`` è ``null`` quando il modello del binding non è nel\ncatalogo: la risoluzione risponde ``MODEL_UNAVAILABLE`` (§19.4).",
        "properties": {
          "binding": {
            "$ref": "#/components/schemas/BindingDTO"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AProfile"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelInfoDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "display_name",
          "version",
          "created_at",
          "updated_at",
          "binding",
          "model"
        ],
        "title": "ProfileDTO",
        "type": "object"
      },
...
      "ResolvedBindingDTO": {
        "description": "Snapshot della risoluzione del binding: riproducibile (ADR 0003).\n\nStesso profilo sullo stesso stato del catalogo → stesso payload.",
        "properties": {
          "binding_id": {
            "format": "uuid",
            "title": "Binding Id",
            "type": "string"
          },
          "binding_name": {
            "title": "Binding Name",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "instructions": {
            "title": "Instructions",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "profile_id": {
            "format": "uuid",
            "title": "Profile Id",
            "type": "string"
          },
          "profile_version": {
            "title": "Profile Version",
            "type": "string"
          },
          "profile_version_id": {
            "format": "uuid",
            "title": "Profile Version Id",
            "type": "string"
          },
          "runtime": {
            "title": "Runtime",
            "type": "string"
          }
        },
        "required": [
          "profile_id",
          "profile_version_id",
          "binding_id",
          "profile_version",
          "binding_name",
          "runtime",
          "model_name",
          "digest",
          "parameters",
          "capabilities",
          "instructions"
        ],
        "title": "ResolvedBindingDTO",
        "type": "object"
      },
      "Role": {
        "description": "Ruolo account del principal (NewRay.md §7.1, 7.2).\n\nAsse separato dal profilo AI: non descrive capacità del modello né\nstrumenti disponibili.",
        "enum": [
          "owner",
          "member",
          "administrator"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "SwitchProfileModelRequest": {
        "description": "Nuova versione del profilo con un modello diverso (B-02.1).\n\n``expected_profile_version`` è l'optimistic lock: il valore\n``version`` dell'ultima lettura del profilo. Se il profilo è\ncambiato nel frattempo la richiesta risponde 409, mai una\nsovrascrittura silenziosa. ``idempotency_key`` opzionale: il retry\ndella stessa richiesta restituisce lo stesso esito senza duplicarlo;\nstessa chiave con payload diverso risponde 409. ``instructions``\nassente mantiene quelle della versione corrente.",
        "properties": {
          "expected_profile_version": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Profile Version"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "instructions": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "model_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "model_name"
        ],
        "title": "SwitchProfileModelRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "NewRay",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/profiles/{profile_id}/binding": {
      "get": {
        "description": "Snapshot riproducibile della risoluzione del binding (ADR 0003).\n\nModello assente dal catalogo → 503 ``MODEL_UNAVAILABLE``\nrecuperabile, mai un fallback invisibile.",
        "operationId": "resolve_binding_api_v1_profiles__profile_id__binding_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedBindingDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Binding"
      }
    },
    "/api/v1/profiles/{profile_id}/versions": {
      "post": {
        "description": "Nuova versione del profilo con un modello diverso (B-02.1).\n\nCrea sempre un binding e una versione nuovi, mai una modifica\nin-place. ``expected_profile_version`` obsoleta o chiave di\nidempotenza riusata con payload diverso → 409. Modello assente\ndal catalogo → 503 ``MODEL_UNAVAILABLE`` recuperabile.",
        "operationId": "switch_model_api_v1_profiles__profile_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchProfileModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switch Model"
      }
    },
    "/api/v1/session": {
      "post": {
        "description": "Bootstrap locale monouso del proprietario (NewRay.md §20.3).\n\nCrea organizzazione, proprietario e sessione, ed emette il cookie\nopaco con HttpOnly/SameSite. Se il proprietario esiste già,\n``CONFLICT`` (409): il bootstrap non può essere ripetuto e non\naccetta il solo nome come credenziale (B-03.2-14).",
        "operationId": "bootstrap_api_v1_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bootstrap"
      }
    },
    "/api/v1/session/login": {
      "post": {
        "description": "Rientro autenticato del proprietario esistente (B-03.2-14).\n\nOgni chiamata riuscita apre una **nuova** sessione: le\nprecedenti restano gestibili con la revoca; nessun riuso del\ncookie in circolo. Errore uniforme ``INVALID_CREDENTIALS`` per\nowner assente o credenziale sbagliata; ``CREDENTIAL_NOT_SET`` è\nil segnale al percorso di recupero locale, non un login riuscito.",
        "operationId": "login_api_v1_session_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
...
  }
}